+++ Challenge/Challenge.VO/EquipmentVO.cs
@@ -8,7 +8,7 @@ namespace Challenge.VO
 {
     public class EquipmentVO
     {
-        public decimal SerialNumber { get; set; }
+        public int SerialNumber { get; set; }
         public string Name { get; set; }
         public byte[] Picture { get; set; }
         public DateTime? NextControlDate { get; set; }
